pyproject.toml
[tool.poetry]
name = "django-flow-integration"
version = "0.1.0"
description = ""
authors = ["Andrew Lee <51726132+hyperleex@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.4"
djangorestframework = "^3.13.1"
httpx = "^0.22.0"
yarl = "^1.7.2"
serializers = "^0.2.4"
djangorestframework-camel-case = "^1.3.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
pytest-django = "^4.5.2"
pytest = "^7.1.1"
pytest-httpx = "^0.20.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
